requirements.txt
albumentations
torch
torchvision
fastapi
uvicorn
pillow
pyyaml
python-multipart